Extract a trait
